change permissions and move files around
