--- src/components/example.tsx
@@ -1,5 +1,0 @@
-const Example = () => {
-  return <div className=""></div>;
-};
-
-export default Example;
