+++ src/components/navbar.tsx
@@ -13,7 +13,7 @@ const Navbar = () => {
     const { error } = await supabase.auth.signInWithOAuth({
       provider: "google",
       options: {
-        redirectTo: `${window.location.origin}/home`,
+        redirectTo: `${window.location.origin}/welcome`,
       },
     });
