+++ src/components/signin.tsx
@@ -8,7 +8,7 @@ const SignIn = () => {
     const { error } = await supabase.auth.signInWithOAuth({
       provider: "google",
       options: {
-        redirectTo: `${window.location.origin}/home`,
+        redirectTo: `${window.location.origin}/welcome`,
       },
     });
